refactor: move helper function from apexUtils.ts to sideBar.ts
